requirements.txt
# Scraping
pyfunda>=2.2.0
funda-scraper>=1.2.0
urllib3<2  # funda-scraper compatibility

# Database
sqlalchemy>=2.0
alembic>=1.13

# Analysis
pandas>=2.1
numpy>=1.26

# Web dashboard
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1

# Configuration
pydantic-settings>=2.1
python-dotenv>=1.0
pyyaml>=6.0

# Scheduling
apscheduler>=3.10

# Utilities
httpx>=0.27
tenacity>=8.2

# Testing
pytest>=7.4
pytest-mock>=3.12
pytest-cov>=4.1
